requirements.txt
fastapi~=0.115.12
pinecone~=7.0.2
opencv-python~=4.11.0.86
python-dotenv~=1.1.0